+++ src/watchdog/observers/__init__.py
@@ -13,8 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-:module: watchdog.observers
+""":module: watchdog.observers
 :synopsis: Observer that picks a native implementation if available.
 :author: yesudeep@google.com (Yesudeep Mangalapilly)
 :author: contact@tiger-222.fr (Mickaël Schoentgen)
@@ -52,47 +51,52 @@ Class          Platforms                        Note
 
 from __future__ import annotations
 
-import sys
+import contextlib
 import warnings
+from typing import TYPE_CHECKING
 
-from watchdog.utils import UnsupportedLibc
+from watchdog.utils import UnsupportedLibc, platform
 
-from .api import BaseObserverSubclassCallable
+if TYPE_CHECKING:
+    from watchdog.observers.api import BaseObserverSubclassCallable
 
-Observer: BaseObserverSubclassCallable
 
+def _get_observer_cls() -> BaseObserverSubclassCallable:
+    if platform.is_linux():
+        with contextlib.suppress(UnsupportedLibc):
+            from watchdog.observers.inotify import InotifyObserver
 
-if sys.platform.startswith("linux"):
-    try:
-        from .inotify import InotifyObserver as Observer
-    except UnsupportedLibc:
-        from .polling import PollingObserver as Observer
-
-elif sys.platform.startswith("darwin"):
-    try:
-        from .fsevents import FSEventsObserver as Observer
-    except Exception:
+            return InotifyObserver
+    elif platform.is_darwin():
         try:
-            from .kqueue import KqueueObserver as Observer
-
-            warnings.warn("Failed to import fsevents. Fall back to kqueue")
+            from watchdog.observers.fsevents import FSEventsObserver
         except Exception:
-            from .polling import PollingObserver as Observer
+            try:
+                from watchdog.observers.kqueue import KqueueObserver
+            except Exception:
+                warnings.warn("Failed to import fsevents and kqueue. Fall back to polling.")
+            else:
+                warnings.warn("Failed to import fsevents. Fall back to kqueue")
+                return KqueueObserver
+        else:
+            return FSEventsObserver
+    elif platform.is_windows():
+        try:
+            from watchdog.observers.read_directory_changes import WindowsApiObserver
+        except Exception:
+            warnings.warn("Failed to import `read_directory_changes`. Fall back to polling.")
+        else:
+            return WindowsApiObserver
+    elif platform.is_bsd():
+        from watchdog.observers.kqueue import KqueueObserver
 
-            warnings.warn("Failed to import fsevents and kqueue. Fall back to polling.")
+        return KqueueObserver
 
-elif sys.platform in ("dragonfly", "freebsd", "netbsd", "openbsd", "bsd"):
-    from .kqueue import KqueueObserver as Observer
+    from watchdog.observers.polling import PollingObserver
 
-elif sys.platform.startswith("win"):
-    try:
-        from .read_directory_changes import WindowsApiObserver as Observer
-    except Exception:
-        from .polling import PollingObserver as Observer
+    return PollingObserver
 
-        warnings.warn("Failed to import read_directory_changes. Fall back to polling.")
 
-else:
-    from .polling import PollingObserver as Observer
+Observer = _get_observer_cls()
 
 __all__ = ["Observer"]