+++ src/watchdog/tricks/__init__.py
@@ -13,8 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-:module: watchdog.tricks
+""":module: watchdog.tricks
 :synopsis: Utility event handlers.
 :author: yesudeep@google.com (Yesudeep Mangalapilly)
 :author: contact@tiger-222.fr (Mickaël Schoentgen)
@@ -41,17 +40,17 @@ Classes
 
 from __future__ import annotations
 
+import contextlib
 import functools
 import logging
 import os
 import signal
 import subprocess
-import sys
 import threading
 import time
 
 from watchdog.events import EVENT_TYPE_OPENED, FileSystemEvent, PatternMatchingEventHandler
-from watchdog.utils import echo
+from watchdog.utils import echo, platform
 from watchdog.utils.event_debouncer import EventDebouncer
 from watchdog.utils.process_watcher import ProcessWatcher
 
@@ -60,7 +59,6 @@ echo_events = functools.partial(echo.echo, write=lambda msg: logger.info(msg))
 
 
 class Trick(PatternMatchingEventHandler):
-
     """Your tricks should subclass this class."""
 
     @classmethod
@@ -80,7 +78,6 @@ class Trick(PatternMatchingEventHandler):
 
 
 class LoggerTrick(Trick):
-
     """A simple trick that does only logs events."""
 
     @echo_events
@@ -89,7 +86,6 @@ class LoggerTrick(Trick):
 
 
 class ShellCommandTrick(Trick):
-
     """Executes shell commands in response to matched events."""
 
     def __init__(
@@ -150,7 +146,8 @@ class ShellCommandTrick(Trick):
             process_watcher = ProcessWatcher(self.process, None)
             self._process_watchers.add(process_watcher)
             process_watcher.process_termination_callback = functools.partial(
-                self._process_watchers.discard, process_watcher
+                self._process_watchers.discard,
+                process_watcher,
             )
             process_watcher.start()
 
@@ -159,7 +156,6 @@ class ShellCommandTrick(Trick):
 
 
 class AutoRestartTrick(Trick):
-
     """Starts a long-running subprocess and restarts it on matched events.
 
     The command parameter is a list of command arguments, such as
@@ -268,11 +264,9 @@ class AutoRestartTrick(Trick):
                             break
                         time.sleep(0.25)
                     else:
-                        try:
+                        # Process is already gone
+                        with contextlib.suppress(OSError):
                             kill_process(self.process.pid, 9)
-                        except OSError:
-                            # Process is already gone
-                            pass
                 self.process = None
         finally:
             self._is_process_stopping = False
@@ -296,7 +290,7 @@ class AutoRestartTrick(Trick):
         self.restart_count += 1
 
 
-if not sys.platform.startswith("win"):
+if not platform.is_windows():
 
     def kill_process(pid, stop_signal):
         os.killpg(os.getpgid(pid), stop_signal)